# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IRolAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponseAllRol": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"totalPage": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IRolAttributes"
						},
						"type": "array"
					},
					"actualPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRolCreationAttributes": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_IRolAttributes.id-or-name_"
					},
					{
						"properties": {
							"permissions": {
								"items": {
									"$ref": "#/components/schemas/IPermissionCreationAttributes"
								},
								"type": "array"
							}
						},
						"type": "object"
					}
				]
			},
...
			"ISurveyAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ESurveyType"
							}
						],
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ESurveyQuestionType": {
				"enum": [
					"selection",
					"text"
				],
				"type": "string"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "api-corpoindustri",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/views/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operación.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"View"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar un usuario por ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de usuario a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rols/show/{rolId}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IRolAttributes"
												}
											],
											"nullable": true
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rol"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "rolId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rols/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": ": IRol[], message?: string }>}",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"anyOf": [
												{
													"items": {
														"$ref": "#/components/schemas/IRolAttributes"
													},
													"type": "array"
												},
												{
													"$ref": "#/components/schemas/IResponseAllRol"
												}
											]
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Obtener todos los datos con paginación.",
				"tags": [
					"Rol"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pag",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rols/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IRolAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rol"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRolCreationAttributes"
							}
						}
					}
				}
			}
		},
...
		"/rols/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar un usuario por ID.",
				"tags": [
					"Rol"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de usuario a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/auth/deleted": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar mi usuario.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/surveys/show/{surveyId}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyAttributes"
												}
											],
											"nullable": true
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Survey"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "surveyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/surveys/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con los datos paginados y un mensaje opcional.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"anyOf": [
												{
													"items": {
														"$ref": "#/components/schemas/ISurveyAttributes"
													},
													"type": "array"
												},
												{
													"$ref": "#/components/schemas/IResponseAllSurvey"
												}
											]
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Obtener todos los datos con paginación.",
				"tags": [
					"Survey"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pag",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/surveys/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Survey"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISurveyCreationAttributes"
							}
						}
					}
				}
			}
		},
		"/surveys/update/{surveyId}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Survey"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "surveyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISurveyAttributes"
							}
						}
					}
				}
			}
		},
		"/surveys/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"Survey"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/surveyResponses/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"SurveyResponse"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/surveyQuestions/show/{surveyQuestionId}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyQuestionAttributes"
												}
											],
											"nullable": true
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"SurveyQuestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "surveyQuestionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/surveyQuestions/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con los datos paginados y un mensaje opcional.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"anyOf": [
												{
													"items": {
														"$ref": "#/components/schemas/ISurveyQuestionAttributes"
													},
													"type": "array"
												},
												{
													"$ref": "#/components/schemas/IResponseAllSurveyQuestion"
												}
											]
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Obtener todos los datos con paginación.",
				"tags": [
					"SurveyQuestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pag",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "text",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ESurveyQuestionType"
						}
					}
				]
			}
		},
		"/surveyQuestions/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyQuestionAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"SurveyQuestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISurveyQuestionCreationAttributes"
							}
						}
					}
				}
			}
		},
		"/surveyQuestions/update/{surveyQuestionId}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/ISurveyQuestionAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"SurveyQuestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "surveyQuestionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISurveyQuestionUpdateAttributes"
							}
						}
					}
				}
			}
		},
		"/surveyQuestions/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"SurveyQuestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/surveyAnswerOptions/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"SurveyAnswerOption"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/suggestions/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"Suggestion"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/products/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"Product"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/favoriteProducts/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la verificacion de la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"FavoriteProduct"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/permissions/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Permission"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/paymentMethods/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"PaymentMethod"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/orders/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/maps/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"Map"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/departments/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"Department"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/conversations/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Promesa que resuelve con la eliminacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una item por ID.",
				"tags": [
					"Conversation"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la item a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}